+++ src/lbm_solver.cpp
@@ -1,6 +1,8 @@
 #include "lbm_solver.h"
 #include "pv.h"
 
+#include <set>
+
 using namespace std;
 
 template <Debug_Level DEBUG_LEVEL>
@@ -43,12 +45,26 @@ void LBMSolver<DEBUG_LEVEL>::initTimers() {
 
 
 template <Debug_Level DEBUG_LEVEL>
+template <GInt NDIM>
 void LBMSolver<DEBUG_LEVEL>::loadConfiguration() {
   m_solverType         = LBSolverType::BGK;  // todo: load from file
   m_method             = LBMethodType::D2Q9; // todo: load from file
   m_noDists            = LBMethod::noDists.at(static_cast<GInt>(m_method));
   m_noVars             = m_dim + 1 + static_cast<GInt>(isThermal());
+  // todo: make settable
   m_outputInfoInterval = 10;
+  m_outputSolutionInterval = 100;
+
+
+  // Re = rho * u * L/nu
+  // u = ma * sqrt(gamma * R * T)
+  //  m_nu = m_ma / gcem::sqrt(3) / m_re * m_refLength;
+  // todo: make settable
+  m_relaxTime = 0.9;
+  m_omega     = 1.0 / m_relaxTime;
+  m_re        = 0.75;
+  m_nu        = (2.0 * m_relaxTime - 1) / 6.0; // default=0.133
+  m_refU      = m_re * m_nu / m_refLength;     // default=1.3333
 
   cerr0 << "<<<<<<<<<<<<>>>>>>>>>>>>>" << std::endl;
   cerr0 << "LBM Type "
@@ -56,6 +72,12 @@ void LBMSolver<DEBUG_LEVEL>::loadConfiguration() {
   cerr0 << "LBM Method "
         << "D2Q9" << std::endl;
   cerr0 << "No. of variables " << std::to_string(m_noVars) << std::endl;
+  cerr0 << "No. Leaf cells: " << grid<NDIM>()->noLeafCells() << std::endl;
+  cerr0 << "No. Bnd cells: " << grid<NDIM>()->noBndCells() << std::endl;
+  cerr0 << "Relaxation Time: " << m_relaxTime << std::endl;
+  cerr0 << "Reynolds Number: " << m_re << std::endl;
+  cerr0 << "Viscosity: " << m_nu << std::endl;
+  cerr0 << "Reference V: " << m_refU << std::endl;
   cerr0 << "+++++++++++++++++++++++++" << std::endl;
 }
 
@@ -76,13 +98,13 @@ template <Debug_Level DEBUG_LEVEL>
 void LBMSolver<DEBUG_LEVEL>::setupMethod() {
   switch(m_method) {
     case LBMethodType::D1Q3:
-      m_weight = {2.0 / 3.0, 1.0 / 6.0, 1.0 / 6.0};
+      m_weight = {1.0 / 6.0, 1.0 / 6.0, 2.0 / 3.0};
       break;
     case LBMethodType::D2Q5:
-      m_weight = {1.0 / 3.0, 1.0 / 6.0, 1.0 / 6.0, 1.0 / 6.0, 1.0 / 6.0};
+      m_weight = {1.0 / 6.0, 1.0 / 6.0, 1.0 / 6.0, 1.0 / 6.0, 1.0 / 3.0};
       break;
     case LBMethodType::D2Q9:
-      m_weight = {4.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 36.0, 1.0 / 36.0, 1.0 / 36.0, 1.0 / 36.0};
+      m_weight = {1.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 36.0, 1.0 / 36.0, 1.0 / 36.0, 1.0 / 36.0, 4.0 / 9.0};
       break;
     default:
       TERMM(-1, "Invalid LBM method type");
@@ -92,8 +114,6 @@ void LBMSolver<DEBUG_LEVEL>::setupMethod() {
 template <Debug_Level DEBUG_LEVEL>
 auto LBMSolver<DEBUG_LEVEL>::run() -> GInt {
   RECORD_TIMER_START(TimeKeeper[Timers::LBMMainLoop]);
-  loadConfiguration();
-  finishInit();
 
   switch(m_dim) {
     case 1:
@@ -118,10 +138,15 @@ template <Debug_Level DEBUG_LEVEL>
 template <GInt NDIM>
 auto LBMSolver<DEBUG_LEVEL>::run() -> GInt {
   RECORD_TIMER_START(TimeKeeper[Timers::LBMMainLoop]);
+  loadConfiguration<NDIM>();
+  finishInit();
   initialCondition<NDIM>();
+  setupPeriodicConnection<NDIM>();
 
-  const GInt noTimesteps = 100;
-  for(GInt ts = 0; ts < noTimesteps; ++ts) {
+  // todo: make settable
+  const GInt noTimesteps = 20000;
+  GBool converged = false;
+  for(GInt ts = 0; ts < noTimesteps && !converged; ++ts) {
     if(m_timeStep > 0 && m_timeStep % m_outputInfoInterval == 0) {
       cerr0 << "Running LBM at " << m_timeStep << std::endl;
       logger << "Running LBM at " << m_timeStep << std::endl;
@@ -146,10 +171,57 @@ auto LBMSolver<DEBUG_LEVEL>::run() -> GInt {
       }
       logger << "rho: " << convRho << " ";
       logger << std::endl;
+
+      // todo: make settable
+      const GDouble maxConv = std::max(std::max(std::max(convUx, convUy), convUz), convRho);
+      if(m_timeStep > 1 && maxConv < 1E-12){
+        logger << "Reached convergence to: " << maxConv <<std::endl;
+        cerr0 << "Reached convergence to: " << maxConv <<std::endl;
+        converged = true;
+      }
     }
     timeStep<NDIM>();
+
+    std::vector<GDouble>              tmp;
+    std::vector<GString>              index;
+    std::vector<std::vector<GString>> values;
+
+    if((m_timeStep > 0 && m_timeStep%m_outputSolutionInterval==0) || m_timeStep == noTimesteps || converged) {
+      // only output leaf cells (i.e. cells without children)
+      std::function<GBool(GInt)> isLeaf = [&](GInt cellId) { return grid<NDIM>()->noChildren(cellId) == 0; };
+
+      tmp.resize(grid<NDIM>()->size());
+      for(GInt cellId = 0; cellId < grid<NDIM>()->size(); ++cellId) {
+        tmp[cellId] = velocity<NDIM>(cellId, 0);
+      }
+
+      index.emplace_back("u");
+      values.emplace_back(toStringVector(tmp, grid<NDIM>()->size()));
+
+      VTK::ASCII::writePoints<NDIM>("test_" + std::to_string(m_timeStep), grid<NDIM>()->size(), grid<NDIM>()->center(), index, values,
+                                    isLeaf);
+    }
+
     ++m_timeStep;
   }
+
+  const GDouble couette_channelHeight = 5.0;
+  auto analytical = [&](const GDouble y){
+    return m_refU/couette_channelHeight * y;
+  };
+
+  GDouble maxError = 0;
+  std::vector<GDouble> error;
+  for(GInt cellId = 0; cellId < grid().size(); ++cellId){
+    const GDouble solution = analytical(grid<NDIM>()->center(cellId, 1));
+    const GDouble delta = velocity<NDIM>(cellId, 0) - solution;
+    maxError = std::max(std::abs(delta), maxError);
+    error.emplace_back(std::sqrt(delta*delta/(solution * solution)));
+  }
+  const GDouble L2error = 1.0/grid().size() * std::accumulate(error.begin(), error.end(), 0.0);
+
+  cerr0 << "max. Error: " << maxError <<std::endl;
+  cerr0 << "avg. L2: " << L2error <<std::endl;
 
   logger << "LBM Solver finished <||" << endl;
   cout << "LBM Solver finished <||" << endl;
@@ -163,7 +235,8 @@ void LBMSolver<DEBUG_LEVEL>::initialCondition() {
   // init to zero:
   for(GInt cellId = 0; cellId < noCells(); ++cellId) {
     for(const auto dir : PV::velocities<NDIM>()) {
-      m_vars[cellId * m_noVars + dir] = 0;
+      m_vars[cellId * m_noVars + dir]    = 0;
+      m_varsold[cellId * m_noVars + dir] = 0;
     }
     rho<NDIM>(cellId) = 1.0;
 
@@ -174,16 +247,6 @@ void LBMSolver<DEBUG_LEVEL>::initialCondition() {
       m_fold[cellId * m_noDists + dist] = m_feq[cellId * m_noDists + dist];
     }
   }
-
-  // Re = rho * u * L/nu
-  // u = ma * sqrt(gamma * R * T)
-  //  m_nu = m_ma / gcem::sqrt(3) / m_re * m_refLength;
-  // todo: make settable
-  m_relaxTime = 0.9;
-  m_omega     = 1.0 / m_relaxTime;
-  m_re        = 10;
-  m_nu        = (2.0 * m_relaxTime - 1) / 6.0; // default=0.133
-  m_refU      = m_re * m_nu / m_refLength;     // default=1.3333
 }
 
 template <Debug_Level DEBUG_LEVEL>
@@ -203,7 +266,7 @@ template <Debug_Level DEBUG_LEVEL>
 template <GInt NDIM>
 void LBMSolver<DEBUG_LEVEL>::currToOldVars() {
   std::copy(m_vars.begin(), m_vars.end(), m_varsold.begin());
-  std::copy(m_f.begin(), m_f.end(), m_fold.begin());
+  //  std::copy(m_f.begin(), m_f.end(), m_fold.begin());
 }
 
 template <Debug_Level DEBUG_LEVEL>
@@ -218,6 +281,7 @@ void LBMSolver<DEBUG_LEVEL>::updateMacroscopicValues() {
                                     + m_fold[cellId * m_noDists + moment[dir][2]] - m_fold[cellId * m_noDists + moment[dir][3]]
                                     - m_fold[cellId * m_noDists + moment[dir][4]] - m_fold[cellId * m_noDists + moment[dir][5]];
     }
+//    cerr0 << "cellId: " << cellId << " (" << velocity<NDIM>(cellId, 0) << ", " << velocity<NDIM>(cellId, 1) << ")" << std::endl;
     rho<NDIM>(cellId) = std::accumulate((m_fold.begin() + cellId * m_noDists), (m_fold.begin() + (cellId + 1) * m_noDists), 0.0);
   }
 }
@@ -241,6 +305,9 @@ void LBMSolver<DEBUG_LEVEL>::collisionStep() {
   for(GInt cellId = 0; cellId < noCells(); ++cellId) {
     for(GInt dist = 0; dist < m_noDists; ++dist) {
       m_f[cellId * m_noDists + dist] = (1 - m_omega) * m_fold[cellId * m_noDists + dist] + m_omega * m_feq[cellId * m_noDists + dist];
+      //      cerr0<< m_f[cellId * m_noDists + dist] << " m_omega " << m_omega << " m_fold " << m_fold[cellId * m_noDists + dist] << " m_feq
+      //      " <<
+      //          m_feq[cellId * m_noDists + dist] <<std::endl;
     }
   }
 }
@@ -249,7 +316,7 @@ template <Debug_Level DEBUG_LEVEL>
 template <GInt NDIM>
 void LBMSolver<DEBUG_LEVEL>::propagationStep() {
   for(GInt cellId = 0; cellId < noCells(); ++cellId) {
-    for(GInt dist = 0; dist < m_noDists; ++dist) {
+    for(GInt dist = 0; dist < m_noDists - 1; ++dist) {
       const GInt neighborId = m_grid->neighbor(cellId, dist);
       if(neighborId != INVALID_CELLID) {
         m_fold[neighborId * m_noDists + dist] = m_f[cellId * m_noDists + dist];
@@ -261,10 +328,39 @@ void LBMSolver<DEBUG_LEVEL>::propagationStep() {
 template <Debug_Level DEBUG_LEVEL>
 template <GInt NDIM>
 void LBMSolver<DEBUG_LEVEL>::boundaryCnd() {
-  cerr0 << "bnd" << std::endl;
+  //  cerr0 << "bnd" << std::endl;
   for(GInt cellId = 0; cellId < noCells(); ++cellId) {
+    //    cerr0 << "bnd " << grid().property(cellId, CellProperties::bndry) << std::endl;
+    //    cerr0 << "leaf " << grid().property(cellId, CellProperties::leaf) << std::endl;
+
     if(grid().property(cellId, CellProperties::bndry) && grid().property(cellId, CellProperties::leaf)) {
-      cerr0 << grid<NDIM>()->center(cellId)[0] << " " << grid<NDIM>()->center(cellId)[1] << std::endl;
+      //      // -X
+      //      if(grid().neighbor(cellId, 0) == INVALID_CELLID) {
+      //        m_fold[cellId * m_noDists + 1] = m_f[cellId * m_noDists + 0];
+      //        m_fold[cellId * m_noDists + 4] = m_f[cellId * m_noDists + 6];
+      //        m_fold[cellId * m_noDists + 5] = m_f[cellId * m_noDists + 7];
+      //      }
+      //      // +X
+      //      if(grid().neighbor(cellId, 1) == INVALID_CELLID) {
+      //        m_fold[cellId * m_noDists + 0] = m_f[cellId * m_noDists + 1];
+      //        m_fold[cellId * m_noDists + 6] = m_f[cellId * m_noDists + 4];
+      //        m_fold[cellId * m_noDists + 7] = m_f[cellId * m_noDists + 5];
+      //      }
+      // -y
+      if(grid().neighbor(cellId, 2) == INVALID_CELLID) {
+        m_fold[cellId * m_noDists + 3] = m_f[cellId * m_noDists + 2];
+        m_fold[cellId * m_noDists + 4] = m_f[cellId * m_noDists + 6];
+        m_fold[cellId * m_noDists + 7] = m_f[cellId * m_noDists + 5];
+      }
+      // +y
+      if(grid().neighbor(cellId, 3) == INVALID_CELLID) {
+        //-y = +y
+        m_fold[cellId * m_noDists + 2] = m_f[cellId * m_noDists + 3];
+        // +x-y = -x+y
+        m_fold[cellId * m_noDists + 5] = m_f[cellId * m_noDists + 7] + 1.0 / 6.0 * (0.1); // todo:testing
+        // +x-y = -x+y
+        m_fold[cellId * m_noDists + 6] = m_f[cellId * m_noDists + 4] - 1.0 / 6.0 * (0.1); // todo:testing
+      }
     }
   }
 }
@@ -325,6 +421,67 @@ auto LBMSolver<DEBUG_LEVEL>::convergence(const GInt var) const -> GDouble {
     conv += std::abs(m_vars[cellId * m_noVars + var] - m_varsold[cellId * m_noVars + var]);
   }
   return conv;
+}
+
+template <Debug_Level DEBUG_LEVEL>
+template <GInt NDIM>
+void LBMSolver<DEBUG_LEVEL>::setupPeriodicConnection() {
+  std::unordered_multimap<GDouble, GInt> coordToCellIds;
+  std::set<GDouble>                      keys;
+  std::set<GInt>                      cellList;
+
+  for(GInt cellId = 0; cellId < noCells(); ++cellId) {
+    if(grid().property(cellId, CellProperties::bndry) && grid().property(cellId, CellProperties::leaf)) {
+      if(grid().neighbor(cellId, 0) == INVALID_CELLID || grid().neighbor(cellId, 1) == INVALID_CELLID) {
+        coordToCellIds.emplace(grid<NDIM>()->center(cellId, 1), cellId);
+        keys.emplace(grid<NDIM>()->center(cellId, 1));
+        cellList.emplace(cellId);
+      }
+    }
+  }
+  for(const auto& p : keys) {
+    if(coordToCellIds.count(p) == 2) {
+      auto search = coordToCellIds.equal_range(p);
+      for(auto it = search.first; it != search.second; ++it) {
+        if(grid().neighbor(it->second, 0) == INVALID_CELLID) {
+          const GInt neighborA                 = it->second;
+          const GInt neighborB                 = (++it)->second;
+          grid<NDIM>()->neighbor(neighborA, 0) = neighborB;
+          grid<NDIM>()->neighbor(neighborB, 1) = neighborA;
+        } else {
+          const GInt neighborA                 = it->second;
+          const GInt neighborB                 = (++it)->second;
+          grid<NDIM>()->neighbor(neighborA, 1) = neighborB;
+          grid<NDIM>()->neighbor(neighborB, 0) = neighborA;
+        }
+
+      }
+    } else {
+      TERMM(-1, "Invalid result!" + std::to_string(p));
+    }
+  }
+
+  //todo: move this periodic code before diagonal detection
+  for(const auto cellId:cellList){
+    const GInt nghbrmX = grid<NDIM>()->neighbor(cellId, 0);
+    const GInt nghbrpX = grid<NDIM>()->neighbor(cellId, 1);
+
+    // +x+y
+    const GInt nghbrpXpY                                 = nghbrpX != INVALID_CELLID ? grid<NDIM>()->neighbor(nghbrpX, 3) : -1;
+    grid<NDIM>()->neighbor(cellId, cartesian::maxNoNghbrs<NDIM>()) = nghbrpXpY;
+
+    // +x-y
+    const GInt nghbrpXmY                                 = nghbrpX != INVALID_CELLID ? grid<NDIM>()->neighbor(nghbrpX, 2) : -1;
+    grid<NDIM>()->neighbor(cellId, cartesian::maxNoNghbrs<NDIM>() + 1) = nghbrpXmY;
+
+    // -x-y
+    const GInt nghbrmXmY                                 = nghbrmX != INVALID_CELLID ? grid<NDIM>()->neighbor(nghbrmX, 2) : -1;
+    grid<NDIM>()->neighbor(cellId, cartesian::maxNoNghbrs<NDIM>() + 2) = nghbrmXmY;
+
+    // -x+y
+    const GInt nghbrmXpY                                 = nghbrmX != INVALID_CELLID ? grid<NDIM>()->neighbor(nghbrmX, 3) : -1;
+    grid<NDIM>()->neighbor(cellId, cartesian::maxNoNghbrs<NDIM>() + 3) = nghbrmXpY;
+  }
 }
 
 template class LBMSolver<Debug_Level::no_debug>;